requirements.txt
google-genai>=0.7.0
pydantic==2.9.2
httpx==0.27.0
requests==2.32.3
tenacity==8.5.0
python-dotenv==1.0.1
pypdf==5.0.1
rapidfuzz==3.9.6
rich==13.9.2
pytest==8.3.3